refactor: move lazyDict to utils.dictionary
