+++ autora/cycle/simple.py
@@ -7,6 +7,7 @@ import numpy as np
 from sklearn.base import BaseEstimator
 
 from autora.experimentalist.pipeline import Pipeline
+from autora.utils.dictionary import LazyDict
 from autora.variable import VariableCollection
 
 
@@ -485,23 +486,6 @@ def _resolve_cycle_properties(params: Dict, cycle_properties: Mapping):
             pass  # no change needed
 
     return params_
-
-
-class LazyDict(Mapping):
-    """Inspired by https://gist.github.com/gyli/9b50bb8537069b4e154fec41a4b5995a"""
-
-    def __init__(self, *args, **kw):
-        self._raw_dict = dict(*args, **kw)
-
-    def __getitem__(self, key):
-        func = self._raw_dict.__getitem__(key)
-        return func()
-
-    def __iter__(self):
-        return iter(self._raw_dict)
-
-    def __len__(self):
-        return len(self._raw_dict)
 
 
 def _get_cycle_properties(data: SimpleCycleData):
